chap08/03.scaling_bilinear.py
import numpy as np, cv2

image = cv2.imread('./images/interpolation.jpg', cv2.IMREAD_GRAYSCALE)
if image is None: raise Exception("영상 파일 읽기 에러")

size = (350, 400)
dst1 = cv2.resize(image, size, 0, 0, cv2.INTER_LINEAR) # OpenCV 함수 적용
dst2 = cv2.resize(image, size, 0, 0, cv2.INTER_NEAREST)

cv2.imshow("image", image)
cv2.imshow("OpenCV_bilinear", dst1)
cv2.imshow("OpenCV_Nearest", dst2)
cv2.waitKey(0)