requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.9.0
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
contourpy==1.3.1
cycler==0.12.1
datascience==0.17.6
decorator==5.1.1
deepface==0.0.93
dlib==19.24.6
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.21
folium==0.19.4
fonttools==4.55.3
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.69.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==24.2
plotly==5.24.1
prompt_toolkit==3.0.50
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
scipy==1.15.1
setuptools==75.1.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.2
xyzservices==2025.1.0
